Add contribute.json guidelines and contribute.dist.json
Added contribute.dist.json, which is a base file for projects’
contribute.json file. It has common fields pre-filled, and examples in
place for other fields.

Also added to README.md the beginnings of structure for adding projects
as well as for how to structure them. In this material, I have details
about what contribute.json is and how to set it up properly for a new
project.